auto-hire-rpa-prototype/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
spacy==3.6.0
scikit-learn==1.3.2
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.0
python-magic==0.4.27
requests==2.32.4
email-validator==2.0.0
SQLAlchemy==2.0.22
alembic==1.11.1
python-dateutil==2.8.2
pytest==8.3.3
Werkzeug==3.0.0
